# execution/64e33dc7/s7l/checkpoints/draft-03-gated/files/docs/contracts/importer-openapi.json
{
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "description": "Machine-readable failure discriminant, present only when the handler set one (e.g. `extension_refresh_invalid`, `invalid`). Absent for generic validation/guard errors.",
            "example": "invalid",
            "type": "string"
          },
          "error": {
            "description": "Short error class name (Nest default), e.g. `Bad Request`, `Unauthorized`.",
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "description": "Human-readable message. A single string for most errors; an array of strings when the global ValidationPipe reports one entry per failed constraint.",
            "example": "Invalid pairing code.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ]
          },
          "path": {
            "description": "Request path the error occurred on.",
            "example": "/api/scout/ingest",
            "type": "string"
          },
          "request_id": {
            "description": "Correlation id from RequestIdMiddleware; omitted when absent.",
            "example": "5f2c9b1e-6d3a-4c8b-9f0e-1a2b3c4d5e6f",
            "type": "string"
          },
          "statusCode": {
            "description": "HTTP status code, echoed in the body.",
            "example": 400,
            "type": "number"
          },
          "timestamp": {
            "description": "ISO-8601 instant the error envelope was built.",
            "example": "2026-07-13T18:30:00.000Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error",
          "timestamp",
          "path"
        ],
        "type": "object"
      },
...
      "RateLimitError": {
        "properties": {
          "error": {
            "description": "Always `Too Many Requests`.",
            "example": "Too Many Requests",
            "type": "string"
          },
          "message": {
            "description": "Generic backoff message (never echoes input or which limit fired).",
            "example": "Too many attempts. Please wait before trying again.",
            "type": "string"
          },
          "retryAfter": {
            "description": "Seconds to wait before retrying; matches the `Retry-After` header.",
            "example": 3600,
            "type": "number"
          },
          "statusCode": {
            "description": "Always 429.",
            "example": 429,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "retryAfter"
        ],
        "type": "object"
      },
...
      "ScoutRunCancelDto": {
        "properties": {
          "intent_id": {
            "description": "Intent id of the run to cancel (the text form of the server setup intent).",
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "intent_id"
        ],
        "type": "object"
      },
      "ScoutRunCancelResult": {
        "properties": {
          "execution_epoch": {
            "description": "Fence epoch after the cancel (strictly greater than before the fence).",
            "minimum": 2,
            "type": "number"
          },
          "intent_id": {
            "description": "The intent id that was cancelled.",
            "type": "string"
          },
          "status": {
            "description": "Always `cancelled`.",
            "enum": [
              "cancelled"
            ],
            "type": "string"
          }
        },
        "required": [
          "intent_id",
          "status",
          "execution_epoch"
        ],
        "type": "object"
      },
      "ScoutRunStartDto": {
        "properties": {
          "import_intent_id": {
            "description": "Server setup intent (ImportIntent.id) to start the server-owned run for. Must be owned by the calling coach, paired and not superseded.",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "import_intent_id"
        ],
        "type": "object"
      },
      "ScoutRunStartResult": {
        "properties": {
          "accepted_start_at": {
            "description": "Server-owned start (set once).",
            "format": "date-time",
            "type": "string"
          },
          "deadline_at": {
            "description": "accepted_start_at + SCOUT_RUN_DEADLINE_MS; enforced lazily, never by a timer.",
            "format": "date-time",
            "type": "string"
          },
          "execution_epoch": {
            "description": "Fence epoch; 1 until the run is fenced.",
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "intent_id": {
            "description": "The run key: the text form of import_intent_id.",
            "type": "string"
          },
          "mode": {
            "description": "Always `server` for a started run.",
            "enum": [
              "server"
            ],
            "type": "string"
          },
          "phase": {
            "description": "Always `discovering` at Start.",
            "enum": [
              "discovering"
            ],
            "type": "string"
          }
        },
        "required": [
          "intent_id",
          "mode",
          "phase",
          "execution_epoch",
          "accepted_start_at",
          "deadline_at"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Supabase-issued JWT. Send as `Authorization: Bearer <token>`. Tokens are verified against the Supabase JWKS in JwtAuthGuard.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Frozen public contract for the tgp-importer Chrome extension + mobile pairing surface. Generated from the backend @nestjs/swagger document — do not hand-edit. Run `npm run contract:importer` to regenerate.",
    "title": "Growth Project API — Importer Contract",
    "version": "2.0.0-c1-s2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scout/runs/cancel": {
      "post": {
        "description": "Fences the run (`fence_reason = cancelled`, epoch + 1) and writes the terminal `cancelled` in one transaction; every in-flight writer on the run is refused from then on. Idempotent: a repeat cancel of a cancelled run returns the same body.",
        "operationId": "ScoutRunController_postCancel",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoutRunCancelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoutRunCancelResult"
                }
              }
            },
            "description": "Run cancelled (or already cancelled)."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed body (global ValidationPipe: whitelist + forbidNonWhitelisted). Standard HttpExceptionFilter envelope; `message` is a string ARRAY of per-field constraint violations when present."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Caller is not a coach or owner."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Uniform not-found: FEATURE_SCOUT_INGEST off (R-DARK-1) OR no run exists for this intent and the calling coach."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    },
                    {
                      "properties": {
                        "code": {
                          "enum": [
                            "run_terminal",
                            "legacy_run"
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "code"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "Lifecycle conflict. `code` is `run_terminal` (another terminal already holds; an open run past its deadline is fenced `timed_out` first) or `legacy_run` (the server does not own client-minted legacy runs and cannot cancel them)."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            },
            "description": "Rate limit exceeded."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an open server-owned run",
        "tags": [
          "scout"
        ]
      }
    },
    "/api/scout/runs/start": {
      "post": {
        "description": "Creates the ONE run row for `import_intent_id` with a server-owned clock: `accepted_start_at` is now and `deadline_at` is now + SCOUT_RUN_DEADLINE_MS (default five minutes), both written once and never updated. Idempotent: a repeat Start for an open run returns the same body. The intent must be owned by the calling coach, paired and not superseded; a run already in a terminal state is not restarted.",
        "operationId": "ScoutRunController_postStart",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoutRunStartDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoutRunStartResult"
                }
              }
            },
            "description": "Run started (or already open)."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Malformed body (global ValidationPipe: whitelist + forbidNonWhitelisted). Standard HttpExceptionFilter envelope; `message` is a string ARRAY of per-field constraint violations when present."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Caller is not a coach or owner."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Uniform not-found: FEATURE_SCOUT_INGEST off (R-DARK-1) OR the intent is unknown to the calling coach. Deliberately indistinguishable — no existence oracle."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    },
                    {
                      "properties": {
                        "code": {
                          "enum": [
                            "intent_not_paired",
                            "intent_superseded",
                            "run_terminal",
                            "legacy_run"
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "code"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "Lifecycle conflict. `code` is one of `intent_not_paired` (pairing incomplete), `intent_superseded` (a newer setup replaced this intent), `run_terminal` (the run already settled; an open run past its deadline is fenced `timed_out` first and answered the same way) or `legacy_run` (the intent names a client-minted legacy run)."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            },
            "description": "Rate limit exceeded."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start the server-owned run for a paired setup intent",
        "tags": [
          "scout"
        ]
      }
    }
  }
}